+++ fotoskazka/app/Models/Service.php
@@ -2,12 +2,15 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Service extends Model
 {
+    use HasFactory;
+
     protected $fillable = [
         'category_id', 'cover_media_id', 'title', 'slug',
         'short_description', 'description', 'price_from',